:construction: refatora SynthInit

+++ VoiceManager.h
@@ -5,11 +5,11 @@
 class VoiceManager
 {
   public:
-    static const int NumberOfVoices = 12;
-    Voice            voices[NumberOfVoices];
-    Voice*           findFreeVoice();
-    void             onNoteOn(int noteNumber);
-    void             onNoteOff(int noteNumber);
+    static const uint8_t NumberOfVoices = 12;
+    Voice                voices[NumberOfVoices];
+    Voice*               findFreeVoice();
+    void                 onNoteOn(int noteNumber);
+    void                 onNoteOff(int noteNumber);
 };
 
 #endif
